move restrict function to middleware
