+++ src/routes/auth.ts
@@ -8,7 +8,7 @@ import {
 } from '../controllers/auth';
 import express from 'express';
 import authMiddleware from '../middleware/authentication';
-import { restrict } from '../controllers/auth';
+import restrict from '../middleware/authorizeRole';
 
 const router = express.Router();
