+++ src/routes/week.ts
@@ -7,7 +7,7 @@ import {
   currentWeek,
 } from '../controllers/week';
 import express from 'express';
-import { restrict } from '../controllers/auth';
+import restrict from '../middleware/authorizeRole';
 
 const router = express.Router();
